File groups can be registered and added tests.
File groups can be registered by calling
Ramaze::Asset::Environment.register_type(). The first parameter is the name of
the type as a symbol (:javascript for example). The second parameter is the
class to use. This class should extend Ramaze::Asset::FileGroup. The type of a
file group is important as it's used in Ramaze::Asset::Environment#serve(). For
example, to serve a Javascript file you'd do the following:

    env = Ramaze::Asset::Environment.new(...)
    env.serve(:javascript, ['...'])

This change makes it possible to add other types such as Less and Coffeescript.

Signed-off-by: YorickPeterse <yorickpeterse@gmail.com>