Fixed a rather serious bug related to file paths.
This bug, which was more or less a missing feature, would cause minified files
to be served relative to the public directory and not relative to the cache
directory. This meant that if the cache directory was "/public/minified" all the
files would be served as "/file" and not "/minified/file".

Signed-off-by: YorickPeterse <yorickpeterse@gmail.com>