Fixed a bug related to not loading processed files
The problem was that the code that checked to see if a file was already loaded
didn't check the controller a file was loaded for. This meant that if you had
two controllers of which both loaded the same file only the first one would be
able to use the file. This small patch fixes this and ensures the code checks
the files for each individual controller.

Signed-off-by: YorickPeterse <yorickpeterse@gmail.com>