Moved the loading of public directories.
Previously Ramaze::Asset::Environment would loop over all the root and public
directories if a new instance of that class was created. This however would
cause problems if a set of paths were added after the instance was created. This
patch changes this behavior by loading the paths in
Ramaze::Asset::Environment#prepare_options() instead of
Ramaze::Asset::Environment#initialize().

Signed-off-by: YorickPeterse <yorickpeterse@gmail.com>